publisher/requirements.txt
kafka-python
zeroconf
websocket-client
